+++ arithmetic/arithmetic_32.hpp
@@ -29,12 +29,12 @@ namespace arithmetic_32 {
         LongDouble(); 
         LongDouble(const LongDouble& x); 
         LongDouble(const LongDouble& x, int precision); 
-        LongDouble(const int &value);
-        LongDouble(const int &value, int precision); 
-        LongDouble(const long long &value);
-        LongDouble(const long long &value, int precision); 
-        LongDouble(const unsigned long long &value); 
-        LongDouble(const unsigned long long &value, int precision); 
+        LongDouble(const int32_t &value);
+        LongDouble(const int32_t &value, int precision); 
+        LongDouble(const int64_t &value);
+        LongDouble(const int64_t &value, int precision); 
+        LongDouble(const uint64_t &value); 
+        LongDouble(const uint64_t &value, int precision); 
         LongDouble(const double &value); 
         LongDouble(const double &value, int precision); 
         LongDouble(const long double &value);
